requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.39
alembic==1.13.1
psycopg[binary]==3.2.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
slowapi==0.1.9
prometheus-client==0.20.0
structlog==24.1.0
httpx==0.27.0
email-validator==2.2.0